+++ libs/designer/src/lib/core/actions/bjsworkflow/operationdeserializer.ts
@@ -15,16 +15,24 @@ import type { NodesMetadata, Operations } from '../../state/workflow/workflowInt
 import type { RootState } from '../../store';
 import { getConnectionReference } from '../../utils/connectors/connections';
 import { isRootNodeInGraph } from '../../utils/graph';
-import { getAllInputParameters, updateDynamicDataInNode, updateTokenMetadata } from '../../utils/parameters/helper';
+import {
+  flattenAndUpdateViewModel,
+  getAllInputParameters,
+  updateDynamicDataInNode,
+  updateTokenMetadata,
+} from '../../utils/parameters/helper';
 import { isTokenValueSegment } from '../../utils/parameters/segment';
 import { initializeOperationDetailsForSwagger } from '../../utils/swagger/operation';
 import { convertOutputsToTokens, getBuiltInTokens, getTokenNodeIds } from '../../utils/tokens';
 import { getAllVariables, getVariableDeclarations, setVariableMetadata } from '../../utils/variables';
-import { getInputParametersFromManifest, getOutputParametersFromManifest, updateCallbackUrlInInputs } from './initialize';
+import {
+  getCustomSwaggerIfNeeded,
+  getInputParametersFromManifest,
+  getOutputParametersFromManifest,
+  updateCallbackUrlInInputs,
+} from './initialize';
 import { getOperationSettings } from './settings';
 import {
-  AppServiceService,
-  ConnectionService,
   LogEntryLevel,
   LoggerService,
   OperationManifestService,
@@ -151,7 +159,13 @@ export const initializeOperationDetailsForManifest = async (
         }
       });
 
-      const { inputs: nodeInputs, dependencies: inputDependencies } = getInputParametersFromManifest(nodeId, manifest, operation);
+      const customSwagger = await getCustomSwaggerIfNeeded(manifest.properties, operation);
+      const { inputs: nodeInputs, dependencies: inputDependencies } = getInputParametersFromManifest(
+        nodeId,
+        manifest,
+        customSwagger,
+        operation
+      );
 
       if (isTrigger) {
         await updateCallbackUrlInInputs(nodeId, nodeOperationInfo, nodeInputs);
@@ -228,6 +242,7 @@ const processChildGraphAndItsInputs = (
             const { inputs: subNodeInputs, dependencies: subNodeInputDependencies } = getInputParametersFromManifest(
               subNodeKey,
               subManifest,
+              /* customSwagger */ undefined,
               subOperation[subNodeKey]
             );
             const subNodeOutputs = { outputs: {} };
@@ -246,6 +261,7 @@ const processChildGraphAndItsInputs = (
         const { inputs: nodeInputs, dependencies: inputDependencies } = getInputParametersFromManifest(
           subGraphKey,
           subManifest,
+          /* customSwagger */ undefined,
           subOperation
         );
         const nodeOutputs = { outputs: {} };
@@ -299,6 +315,20 @@ const updateTokenMetadataInParameters = (
 
           return segment;
         });
+      }
+
+      const viewModel = parameter.editorViewModel;
+      if (viewModel) {
+        flattenAndUpdateViewModel(
+          viewModel,
+          actionNodes,
+          triggerNodeId,
+          nodesData,
+          operations,
+          workflowParameters,
+          nodesMetadata,
+          parameter.type
+        );
       }
     }
   }
@@ -418,50 +448,4 @@ export const updateDynamicDataInNodes = async (getState: () => RootState, dispat
       operation
     );
   }
-};
-
-export const initializeSwaggerDynamicData = async (workflow: DeserializedWorkflow): Promise<any> => {
-  await Promise.all(
-    Object.entries(workflow.actionData).map(async ([id, operation]) => {
-      const newOperation = await initializeOperationSwaggerDynamicData(operation);
-      if (newOperation) {
-        workflow.actionData[id] = {
-          ...operation,
-          ...newOperation,
-        };
-      }
-    })
-  );
-};
-
-const initializeOperationSwaggerDynamicData = async (operation: any): Promise<LogicAppsV2.OperationDefinition | undefined> => {
-  const { inputs, metadata } = operation;
-  if (inputs?.operationId) return;
-
-  const { apiDefinitionUrl, swaggerSource } = metadata ?? {};
-  if (!apiDefinitionUrl || (swaggerSource !== 'custom' && swaggerSource !== 'website')) return;
-
-  const { uri, method } = inputs ?? {};
-  if (!uri || !method) return;
-
-  let swaggerOperation;
-  if (swaggerSource === 'custom') {
-    const connectionService = ConnectionService();
-    swaggerOperation = await connectionService.getOperationFromPathAndMethod(apiDefinitionUrl, uri, method);
-  } else if (swaggerSource === 'website') {
-    const appServiceService = AppServiceService();
-    swaggerOperation = await appServiceService.getOperationFromPathAndMethod(apiDefinitionUrl, uri, method);
-  }
-  if (!swaggerOperation) return;
-  const { operationId } = swaggerOperation;
-
-  operation = {
-    ...operation,
-    inputs: {
-      ...operation.inputs,
-      operationId,
-    },
-  };
-
-  return operation;
 };