+++ libs/designer/src/lib/core/queries/connector.ts
@@ -43,12 +43,10 @@ export const getListDynamicValues = async (
   parameterAlias: string | undefined,
   parameters: Record<string, any>,
   dynamicState: any,
-  nodeInputs: any,
   nodeMetadata: any
 ): Promise<ListDynamicValue[]> => {
   const queryClient = getReactQueryClient();
   const service = ConnectorService();
-  const parameterKey = getParametersKey(parameters);
 
   return queryClient.fetchQuery(
     [
@@ -56,19 +54,9 @@ export const getListDynamicValues = async (
       (connectionId ?? '').toLowerCase(),
       connectorId.toLowerCase(),
       operationId.toLowerCase(),
-      parameterKey ? parameterKey.toLowerCase() : parameterAlias,
+      getParametersKey(parameters).toLowerCase(),
     ],
-    () =>
-      service.getListDynamicValues(
-        connectionId,
-        connectorId,
-        operationId,
-        parameterAlias,
-        parameters,
-        dynamicState,
-        nodeInputs,
-        nodeMetadata
-      )
+    () => service.getListDynamicValues(connectionId, connectorId, operationId, parameterAlias, parameters, dynamicState, nodeMetadata)
   );
 };
 
@@ -110,7 +98,6 @@ export const getDynamicSchemaProperties = async (
   parameterAlias: string | undefined,
   parameters: Record<string, any>,
   dynamicState: any,
-  nodeInputs: any,
   nodeMetadata: any
 ): Promise<OpenAPIV2.SchemaObject> => {
   const queryClient = getReactQueryClient();
@@ -125,8 +112,7 @@ export const getDynamicSchemaProperties = async (
       getParametersKey(parameters),
       `isInput:${!!dynamicState?.isInput}`,
     ],
-    () =>
-      service.getDynamicSchema(connectionId, connectorId, operationId, parameterAlias, parameters, dynamicState, nodeInputs, nodeMetadata)
+    () => service.getDynamicSchema(connectionId, connectorId, operationId, parameterAlias, parameters, dynamicState, nodeMetadata)
   );
 };