+++ libs/designer/src/lib/core/utils/parameters/helper.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-case-declarations  */
 import constants from '../../../common/constants';
 import type { ConnectionReference, WorkflowParameter } from '../../../common/models/workflow';
 import type { NodeDataWithOperationMetadata } from '../../actions/bjsworkflow/operationdeserializer';
@@ -29,11 +30,14 @@ import { getParentArrayKey, isForeachActionNameForLoopsource } from '../loops';
 import { isOneOf } from '../openapi/schema';
 import { loadDynamicOutputsInNode } from '../outputs';
 import { hasSecureOutputs } from '../setting';
+import { processPathInputs } from '../swagger/inputsbuilder';
+import { extractPathFromUri, getOperationIdFromDefinition } from '../swagger/operation';
 import { convertWorkflowParameterTypeToSwaggerType } from '../tokens';
 import { validateJSONParameter, validateStaticParameterInfo } from '../validation';
 import { getRecurrenceParameters } from './builtins';
 import { addCastToExpression, addFoldingCastToExpression } from './casting';
 import { getDynamicInputsFromSchema, getDynamicSchema, getDynamicValues } from './dynamicdata';
+import { requiresFilePickerEditor } from './picker';
 import {
   createLiteralValueSegment,
   isExpressionToken,
@@ -45,6 +49,7 @@ import {
   isOutputTokenValueSegment,
   isParameterToken,
   isTokenValueSegment,
+  isValueSegment,
   isVariableToken,
   ValueSegmentConvertor,
 } from './segment';
@@ -87,7 +92,7 @@ import type {
   SwaggerParser,
 } from '@microsoft/parsers-logic-apps';
 import {
-  DeserializationLocation,
+  DeserializationType,
   PropertySerializationType,
   getKnownTitles,
   isLegacyDynamicValuesExtension,
@@ -261,8 +266,7 @@ export function toParameterInfoMap(inputParameters: InputParameter[], stepDefini
 
 /**
  * Gets the parameter info object for UI elements from the resolved parameters from schema, swagger, definition, etc.
- * @arg {InputParameter} parameter - An object with metadata about a Swagger input parameter.
- * @arg {RepetitionContext} repetitionContext - An object contains the repetition related context data.
+ * @arg {ResolvedParameter} parameter - An object with metadata about a Swagger input parameter.
  * @arg {Record<string, string>} [metadata] - A hash mapping dynamic value lookup values to their display strings.
  * @arg {boolean} [shouldIgnoreDefaultValue=false] - True if should not populate with default value of dynamic parameter.
  * @return {ParameterInfo} - An object with the view model for an input parameter field.
@@ -272,9 +276,10 @@ export function createParameterInfo(
   _metadata?: Record<string, string>,
   shouldIgnoreDefaultValue = false
 ): ParameterInfo {
-  const { editor, editorOptions, editorViewModel, schema } = getParameterEditorProps(parameter, shouldIgnoreDefaultValue);
+  const isFilePicker = requiresFilePickerEditor(parameter);
+  const { editor, editorOptions, editorViewModel, schema } = getParameterEditorProps(parameter, shouldIgnoreDefaultValue, isFilePicker);
   const value = loadParameterValue(parameter);
-  const { alias, dependencies, encode, format, isDynamic, isUnknown, serialization } = parameter;
+  const { alias, dependencies, encode, format, isDynamic, isUnknown, serialization, deserialization } = parameter;
   const info = {
     alias,
     dependencies,
@@ -284,6 +289,7 @@ export function createParameterInfo(
     isDynamic: !!isDynamic,
     isUnknown,
     serialization,
+    deserialization,
   };
 
   const parameterInfo: ParameterInfo = {
@@ -333,8 +339,12 @@ function hasValue(parameter: ResolvedParameter): boolean {
 }
 
 // TODO - Need to figure out a way to get the managedIdentity for the app for authentication editor
-export function getParameterEditorProps(parameter: InputParameter, shouldIgnoreDefaultValue = false): ParameterEditorProps {
-  const { dynamicValues, type, itemSchema, visibility, value } = parameter;
+export function getParameterEditorProps(
+  parameter: InputParameter,
+  shouldIgnoreDefaultValue = false,
+  isFilePicker?: boolean
+): ParameterEditorProps {
+  const { dynamicValues, type, itemSchema, visibility, value, enum: schemaEnum } = parameter;
   let { editor, editorOptions, schema } = parameter;
   let editorViewModel;
   if (!editor) {
@@ -342,12 +352,14 @@ export function getParameterEditorProps(parameter: InputParameter, shouldIgnoreD
       editor = constants.EDITOR.ARRAY;
       editorViewModel = initializeArrayViewModel(parameter, shouldIgnoreDefaultValue);
       schema = { ...schema, ...{ 'x-ms-editor': editor } };
-    } else if (schema && (schema.enum || schema[ExtensionProperties.CustomEnum]) && !equals(visibility, Visibility.Internal)) {
+    } else if (schemaEnum || schema?.enum || (schema?.[ExtensionProperties.CustomEnum] && !equals(visibility, Visibility.Internal))) {
       editor = constants.EDITOR.COMBOBOX;
       schema = { ...schema, ...{ 'x-ms-editor': editor } };
 
       let schemaEnumOptions: ComboboxItem[];
-      if (schema[ExtensionProperties.CustomEnum]) {
+      if (schemaEnum) {
+        schemaEnumOptions = schemaEnum.map((enumItem) => ({ ...enumItem, key: enumItem.value?.toString() }));
+      } else if (schema[ExtensionProperties.CustomEnum]) {
         schemaEnumOptions = schema[ExtensionProperties.CustomEnum];
       } else {
         schemaEnumOptions = schema.enum.map(
@@ -377,13 +389,47 @@ export function getParameterEditorProps(parameter: InputParameter, shouldIgnoreD
     editorViewModel = toAuthenticationViewModel(value);
     editorOptions = { ...editorOptions, identity: WorkflowService().getAppIdentity?.() };
   } else if (editor === constants.EDITOR.CONDITION) {
-    editorViewModel = editorOptions?.isOldFormat ? toUntilViewModel(value) : toConditionViewModel(value);
+    editorViewModel = editorOptions?.isOldFormat ? toSimpleQueryBuilderViewModel(value) : toConditionViewModel(value);
   } else if (dynamicValues && isLegacyDynamicValuesExtension(dynamicValues) && dynamicValues.extension.builtInOperation) {
     editor = undefined;
+  } else if (isFilePicker) {
+    editor = constants.EDITOR.FILEPICKER;
   }
   return { editor, editorOptions, editorViewModel, schema };
 }
 
+// Helper Functions for Creating Editor View Models
+const containsExpression = (operand: string): boolean => {
+  return operand.includes('(') && operand.includes(')');
+};
+
+const convertStringToInputParameter = (
+  value: string,
+  removeQuotesFromExpression?: boolean,
+  trimExpression?: boolean,
+  convertIfContainsExpression?: boolean
+): InputParameter => {
+  const hasExpression = containsExpression(value);
+  let newValue = value;
+  if (removeQuotesFromExpression) {
+    newValue = removeQuotes(newValue);
+  }
+  if (trimExpression) {
+    newValue = newValue.trim();
+  }
+  if (hasExpression && convertIfContainsExpression && !newValue.startsWith('@')) {
+    newValue = `@${newValue}`;
+  }
+  return {
+    key: guid(),
+    name: newValue,
+    type: 'any',
+    hideInUI: false,
+    value: newValue,
+  };
+};
+
+// Create Array Editor View Model
 const toArrayViewModel = (input: any): { schema: any } => {
   const schema: any = destructureSchema(input.itemSchema);
   return { schema };
@@ -406,18 +452,16 @@ const destructureSchema = (schema: any): any => {
   return newSchema;
 };
 
-const toUntilViewModel = (input: any): { isOldFormat: boolean; items: RowItemProps } => {
+// Create SimpleQueryBuilder Editor View Model
+const toSimpleQueryBuilderViewModel = (input: any): { isOldFormat: boolean; items: RowItemProps } => {
   let operand1: ValueSegment[], operand2: ValueSegment[], operation: string;
   try {
     operation = input.substring(input.indexOf('@') + 1, input.indexOf('('));
     const operations = input.split(',');
-    operand1 = loadParameterValue(
-      convertStringToInputParameter(removeQuotes(operations[0].substring(operations[0].indexOf('(') + 1)).trim())
-    );
-
-    operand2 = loadParameterValue(
-      convertStringToInputParameter(removeQuotes(operations[1].substring(0, operations[1].indexOf(')'))).trim())
-    );
+    const operand1String = operations[0].substring(operations[0].indexOf('(') + 1);
+    const operand2String = operations[1].substring(0, operations[1].lastIndexOf(')'));
+    operand1 = loadParameterValue(convertStringToInputParameter(operand1String, true, true, true));
+    operand2 = loadParameterValue(convertStringToInputParameter(operand2String, true, true, true));
   } catch {
     operation = 'equals';
     operand1 = [];
@@ -430,16 +474,7 @@ const toUntilViewModel = (input: any): { isOldFormat: boolean; items: RowItemPro
   };
 };
 
-const convertStringToInputParameter = (s: string): InputParameter => {
-  return {
-    key: s,
-    name: s,
-    type: 'any',
-    hideInUI: false,
-    value: s,
-  };
-};
-
+// Create QueryBuilder Editor View Model
 export const toConditionViewModel = (input: any): { items: GroupItemProps } => {
   const getConditionOption = getConditionalSelectedOption(input);
   const items: GroupItemProps = {
@@ -477,8 +512,12 @@ function recurseConditionalItems(input: any, selectedOption?: GroupDropdownOptio
         output.push({
           type: GroupType.ROW,
           operator: not + dropdownVal,
-          operand1: loadParameterValue({ value: not ? item[not][dropdownVal][0] : item[dropdownVal][0] } as InputParameter),
-          operand2: loadParameterValue({ value: not ? item[not][dropdownVal][1] : item[dropdownVal][1] } as InputParameter),
+          operand1: loadParameterValue(
+            convertStringToInputParameter(not ? item[not][dropdownVal][0] : item[dropdownVal][0], true, true, true)
+          ),
+          operand2: loadParameterValue(
+            convertStringToInputParameter(not ? item[not][dropdownVal][1] : item[dropdownVal][1], true, true, true)
+          ),
         });
       } else {
         output.push({ type: GroupType.GROUP, condition: condition, items: recurseConditionalItems(item, condition) });
@@ -488,6 +527,7 @@ function recurseConditionalItems(input: any, selectedOption?: GroupDropdownOptio
   return output;
 }
 
+// Create Dictionary Editor View Model
 function toDictionaryViewModel(value: any): { items: DictionaryEditorItemProps[] | undefined } {
   let items: DictionaryEditorItemProps[] | undefined = [];
   const valueToParse = value !== null ? value ?? {} : value;
@@ -497,8 +537,8 @@ function toDictionaryViewModel(value: any): { items: DictionaryEditorItemProps[]
     const keys = Object.keys(valueToParse);
     for (const itemKey of keys) {
       items.push({
-        key: loadParameterValue({ value: itemKey } as any),
-        value: loadParameterValue({ value: valueToParse[itemKey] } as any),
+        key: loadParameterValue(convertStringToInputParameter(itemKey)),
+        value: loadParameterValue(convertStringToInputParameter(valueToParse[itemKey])),
       });
     }
 
@@ -512,6 +552,7 @@ function toDictionaryViewModel(value: any): { items: DictionaryEditorItemProps[]
   return { items };
 }
 
+// Create Table Editor View Model
 function toTableViewModel(value: any, editorOptions: any): { items: DictionaryEditorItemProps[]; columnMode: ColumnMode } {
   const placeholderItem = { key: [createLiteralValueSegment('')], value: [createLiteralValueSegment('')] };
   if (Array.isArray(value)) {
@@ -519,8 +560,8 @@ function toTableViewModel(value: any, editorOptions: any): { items: DictionaryEd
     const items: DictionaryEditorItemProps[] = [];
     for (const item of value) {
       items.push({
-        key: loadParameterValue({ value: item[keys[0]] } as any),
-        value: loadParameterValue({ value: item[keys[1]] } as any),
+        key: loadParameterValue(convertStringToInputParameter(item[keys[0]])),
+        value: loadParameterValue(convertStringToInputParameter(item[keys[1]])),
       });
     }
 
@@ -530,6 +571,7 @@ function toTableViewModel(value: any, editorOptions: any): { items: DictionaryEd
   return { items: [placeholderItem], columnMode: ColumnMode.Automatic };
 }
 
+// Create Authentication Editor View Model
 function toAuthenticationViewModel(value: any): { type: AuthenticationType; authenticationValue: AuthProps } {
   const emptyValue = { type: AuthenticationType.NONE, authenticationValue: {} };
 
@@ -540,8 +582,8 @@ function toAuthenticationViewModel(value: any): { type: AuthenticationType; auth
           type: value.type,
           authenticationValue: {
             basic: {
-              basicUsername: loadParameterValue({ value: value.username } as InputParameter),
-              basicPassword: loadParameterValue({ value: value.password } as InputParameter),
+              basicUsername: loadParameterValue(convertStringToInputParameter(value.username)),
+              basicPassword: loadParameterValue(convertStringToInputParameter(value.password)),
             },
           },
         };
@@ -550,8 +592,8 @@ function toAuthenticationViewModel(value: any): { type: AuthenticationType; auth
           type: value.type,
           authenticationValue: {
             clientCertificate: {
-              clientCertificatePfx: loadParameterValue({ value: value.pfx } as InputParameter),
-              clientCertificatePassword: loadParameterValue({ value: value.password } as InputParameter),
+              clientCertificatePfx: loadParameterValue(convertStringToInputParameter(value.pfx)),
+              clientCertificatePassword: loadParameterValue(convertStringToInputParameter(value.password)),
             },
           },
         };
@@ -561,12 +603,12 @@ function toAuthenticationViewModel(value: any): { type: AuthenticationType; auth
           type: value.type,
           authenticationValue: {
             aadOAuth: {
-              oauthTenant: loadParameterValue({ value: value.tenant } as InputParameter),
-              oauthAudience: loadParameterValue({ value: value.audience } as InputParameter),
-              oauthClientId: loadParameterValue({ value: value.clientId } as InputParameter),
-              oauthTypeSecret: loadParameterValue({ value: value.secret } as InputParameter),
-              oauthTypeCertificatePfx: loadParameterValue({ value: value.pfx } as InputParameter),
-              oauthTypeCertificatePassword: loadParameterValue({ value: value.password } as InputParameter),
+              oauthTenant: loadParameterValue(convertStringToInputParameter(value.tenant)),
+              oauthAudience: loadParameterValue(convertStringToInputParameter(value.audience)),
+              oauthClientId: loadParameterValue(convertStringToInputParameter(value.clientId)),
+              oauthTypeSecret: loadParameterValue(convertStringToInputParameter(value.secret)),
+              oauthTypeCertificatePfx: loadParameterValue(convertStringToInputParameter(value.pfx)),
+              oauthTypeCertificatePassword: loadParameterValue(convertStringToInputParameter(value.password)),
             },
           },
         };
@@ -576,7 +618,7 @@ function toAuthenticationViewModel(value: any): { type: AuthenticationType; auth
           type: value.type,
           authenticationValue: {
             raw: {
-              rawValue: loadParameterValue({ value: value.value } as InputParameter),
+              rawValue: loadParameterValue(convertStringToInputParameter(value.value)),
             },
           },
         };
@@ -586,7 +628,7 @@ function toAuthenticationViewModel(value: any): { type: AuthenticationType; auth
           type: value.type,
           authenticationValue: {
             msi: {
-              msiAudience: loadParameterValue({ value: value.audience } as InputParameter),
+              msiAudience: loadParameterValue(convertStringToInputParameter(value.audience)),
               msiIdentity: value.identity,
             },
           },
@@ -1782,14 +1824,14 @@ function getStringifiedValueFromEditorViewModel(parameter: ParameterInfo, isDefi
       return undefined;
     case constants.EDITOR.CONDITION:
       return editorOptions?.isOldFormat
-        ? serializeUntil(editorViewModel)
+        ? serializeSimpleQueryBuilder(editorViewModel)
         : JSON.stringify(recurseSerializeCondition(parameter, editorViewModel.items, isDefinitionValue));
     default:
       return undefined;
   }
 }
 
-export const serializeUntil = (editorViewModel: any): string => {
+export const serializeSimpleQueryBuilder = (editorViewModel: any): string => {
   return editorViewModel.value;
 };
 
@@ -1918,6 +1960,7 @@ export function getGroupAndParameterFromParameterKey(
 export function getInputsValueFromDefinitionForManifest(
   inputsLocation: string[],
   manifest: OperationManifest,
+  customSwagger: SwaggerParser | undefined,
   stepDefinition: any,
   allInputs: InputParameter[]
 ): any {
@@ -1931,7 +1974,7 @@ export function getInputsValueFromDefinitionForManifest(
 
   inputsValue = swapInputsValueIfNeeded(inputsValue, manifest);
 
-  return updateInputsValueForSpecialCases(inputsValue, allInputs);
+  return updateInputsValueForSpecialCases(inputsValue, allInputs, customSwagger);
 }
 
 function swapInputsValueIfNeeded(inputsValue: any, manifest: OperationManifest) {
@@ -1969,7 +2012,7 @@ function swapInputsValueIfNeeded(inputsValue: any, manifest: OperationManifest) 
  * serialization or special cases for deserialiation like reading values from inputs for non serializable parameter.
  * Example: Batch trigger
  */
-function updateInputsValueForSpecialCases(inputsValue: any, allInputs: InputParameter[]) {
+function updateInputsValueForSpecialCases(inputsValue: any, allInputs: InputParameter[], customSwagger?: SwaggerParser) {
   if (!inputsValue || typeof inputsValue !== 'object') {
     return inputsValue;
   }
@@ -1979,39 +2022,78 @@ function updateInputsValueForSpecialCases(inputsValue: any, allInputs: InputPara
 
   for (const propertyParameter of propertyNameParameters) {
     const { name, serialization } = propertyParameter;
-    if (serialization?.property?.type === PropertySerializationType.ParentObject) {
-      const parameterReference = serialization.property.parameterReference;
-      const parentPropertyName = parameterReference.substring(0, parameterReference.lastIndexOf('.'));
-      const parentPropertySegments = parentPropertyName.split('.');
-      const objectValue = getObjectPropertyValue(finalValue, parentPropertySegments);
+    if (serialization?.property) {
+      const parameterReference = serialization.property.parameterReference as string;
 
-      // NOTE: Currently this only handles only one variable property name in the object
-      const keys = Object.keys(objectValue);
-      if (keys.length !== 1) {
-        return inputsValue;
+      switch (serialization?.property?.type?.toLowerCase()) {
+        case PropertySerializationType.ParentObject:
+          const parentPropertyName = parameterReference.substring(0, parameterReference.lastIndexOf('.'));
+          const parentPropertySegments = parentPropertyName.split('.');
+          const objectValue = getObjectPropertyValue(finalValue, parentPropertySegments);
+
+          // NOTE: Currently this only handles only one variable property name in the object
+          const keys = Object.keys(objectValue);
+          if (keys.length !== 1) {
+            return inputsValue;
+          }
+
+          const propertyName = keys[0];
+
+          safeSetObjectPropertyValue(finalValue, parentPropertySegments, {
+            [serialization.property.name as string]: {
+              ...objectValue[propertyName],
+              [name.substring(name.lastIndexOf('.') + 1)]: propertyName,
+            },
+          });
+          break;
+
+        case PropertySerializationType.PathTemplate:
+          const parameterLocation = propertyParameter.name.split('.');
+          const pathParametersLocation = parameterReference.split('.');
+          const uriValue = getObjectPropertyValue(finalValue, parameterLocation);
+          const pathParameters = processPathInputs(uriValue, propertyParameter.default);
+          safeSetObjectPropertyValue(finalValue, pathParametersLocation, pathParameters);
+          safeSetObjectPropertyValue(finalValue, parameterLocation, propertyParameter.default);
+          break;
+
+        default:
+          break;
       }
-
-      const propertyName = keys[0];
-
-      safeSetObjectPropertyValue(finalValue, parentPropertySegments, {
-        [serialization.property.name]: {
-          ...objectValue[propertyName],
-          [name.substring(name.lastIndexOf('.') + 1)]: propertyName,
-        },
-      });
     }
   }
 
+  // NOTE: Deserialization should be processed only after parameters serialization property is processed.
   const parametersWithDeserialization = allInputs.filter((input) => !!input.deserialization);
   for (const parameter of parametersWithDeserialization) {
     const { name, deserialization } = parameter;
-    if (deserialization?.type === DeserializationLocation.ParentObjectProperties) {
-      const parentPropertySegments = deserialization.parameterReference.split('.');
-      const objectValue = getObjectPropertyValue(finalValue, parentPropertySegments);
-      const value = Object.keys(objectValue ?? {});
+    if (deserialization?.parameterReference) {
+      const { type, parameterReference, options } = deserialization;
+      const propertySegments = parameterReference.split('.');
 
-      objectValue[name.split('.').at(-1) as string] = value;
-      safeSetObjectPropertyValue(finalValue, parentPropertySegments, objectValue);
+      switch (type) {
+        case DeserializationType.ParentObjectProperties:
+          const objectValue = getObjectPropertyValue(finalValue, propertySegments);
+          const value = Object.keys(objectValue ?? {});
+          objectValue[name.split('.').at(-1) as string] = value;
+          safeSetObjectPropertyValue(finalValue, propertySegments, objectValue);
+          break;
+
+        case DeserializationType.SwaggerOperationId:
+          const method = getObjectPropertyValue(finalValue, options?.swaggerOperation.methodPath as string[]);
+          const uri = getObjectPropertyValue(finalValue, options?.swaggerOperation.uriPath as string[]);
+          const operationId = getOperationIdFromDefinition(
+            {
+              method,
+              path: extractPathFromUri(uri, customSwagger?.api.basePath as string),
+            },
+            customSwagger as SwaggerParser
+          );
+          safeSetObjectPropertyValue(finalValue, propertySegments, operationId);
+          break;
+
+        default:
+          break;
+      }
     }
   }
 
@@ -2107,8 +2189,69 @@ export function updateTokenMetadataInParameters(parameters: ParameterInfo[], roo
         return segment;
       });
     }
+    const viewModel = parameter.editorViewModel;
+    if (viewModel) {
+      flattenAndUpdateViewModel(viewModel, actionNodes, triggerNodeId, nodesData, operations, definitions, nodesMetadata, parameter.type);
+    }
   }
 }
+
+export const flattenAndUpdateViewModel = (
+  items: any,
+  actionNodes: Record<string, string>,
+  triggerNodeId: string,
+  nodes: Record<string, Partial<NodeDataWithOperationMetadata>>,
+  operations: Actions,
+  workflowParameters: Record<string, WorkflowParameter | WorkflowParameterDefinition>,
+  nodesMetadata: NodesMetadata,
+  parameterType?: string
+) => {
+  if (!items) return;
+  // check if on bottom-most level (array of valueSegments)
+  if (Array.isArray(items) && items.every((item) => isValueSegment(item))) {
+    return items.map((keyItem: ValueSegment) => {
+      if (!isTokenValueSegment(keyItem)) return keyItem;
+      const valueSegmentToUpdate = updateTokenMetadata(
+        keyItem,
+        actionNodes,
+        triggerNodeId,
+        nodes,
+        operations,
+        workflowParameters,
+        nodesMetadata,
+        parameterType
+      );
+      if (valueSegmentToUpdate) return valueSegmentToUpdate;
+      return keyItem;
+    }) as ValueSegment[];
+  }
+
+  const replacedItems: any = {};
+  Object.entries(items).forEach(([itemKey, itemValue]) => {
+    if (typeof itemValue === 'object') {
+      replacedItems[itemKey] = flattenAndUpdateViewModel(
+        itemValue,
+        actionNodes,
+        triggerNodeId,
+        nodes,
+        operations,
+        workflowParameters,
+        nodesMetadata,
+        parameterType
+      );
+    } else {
+      replacedItems[itemKey] = itemValue;
+    }
+  });
+  // Keep Array Format and preserve the order of the elements
+  return Array.isArray(items)
+    ? Object.keys(replacedItems)
+        .map(Number)
+        .sort((a, b) => a - b)
+        .map((key) => replacedItems[key])
+    : replacedItems;
+};
+
 export function updateTokenMetadata(
   valueSegment: ValueSegment,
   actionNodes: Record<string, string>,