+++ libs/services/designer-client-services/src/lib/base/operationmanifest.ts
@@ -43,6 +43,7 @@ import { ArgumentException, equals, UnsupportedException } from '@microsoft/util
 import type { OperationInfo, OperationManifest } from '@microsoft/utils-logic-apps';
 
 const apimanagement = 'apimanagement';
+const apimanagementtrigger = 'apimanagementtrigger';
 const as2Encode = 'as2encode';
 const as2Decode = 'as2decode';
 const integrationaccountartifactlookup = 'integrationaccountartifactlookup';
@@ -128,6 +129,7 @@ const dataMapperConnectorId = 'connectionProviders/dataMapperOperations';
 
 const azurefunction = 'azurefunction';
 const appservice = 'appservice';
+const appservicetrigger = 'appservicetrigger';
 const invokeworkflow = 'invokeworkflow';
 
 export const supportedBaseManifestTypes = [
@@ -314,7 +316,7 @@ export function getBuiltInOperationInfo(definition: any, isTrigger: boolean): Op
       if (equals(definition?.metadata?.swaggerSource, 'website')) {
         return {
           connectorId: appServiceConnectorId,
-          operationId: appservice,
+          operationId: isTrigger ? appservicetrigger : appservice,
         };
       } else {
         return {
@@ -403,19 +405,13 @@ export function getBuiltInOperationInfo(definition: any, isTrigger: boolean): Op
     case apimanagement:
       return {
         connectorId: apiManagementConnectorId,
-        operationId: apimanagement,
-      };
-
-    case sendtobatch:
-      return {
-        connectorId: batchConnectorId,
-        operationId: sendtobatch,
+        operationId: isTrigger ? apimanagementtrigger : apimanagement,
       };
 
     case appservice:
       return {
         connectorId: appServiceConnectorId,
-        operationId: appservice,
+        operationId: isTrigger ? appservicetrigger : appservice,
       };
 
     case azurefunction:
@@ -453,10 +449,6 @@ export function getBuiltInOperationInfo(definition: any, isTrigger: boolean): Op
 }
 
 const builtInOperationsMetadata: Record<string, OperationInfo> = {
-  [apimanagement]: {
-    connectorId: apiManagementConnectorId,
-    operationId: 'apiManagement',
-  },
   [appendtoarrayvariable]: {
     connectorId: variableConnectorId,
     operationId: appendtoarrayvariable,
@@ -476,6 +468,10 @@ const builtInOperationsMetadata: Record<string, OperationInfo> = {
   [batch]: {
     connectorId: batchConnectorId,
     operationId: batch,
+  },
+  [sendtobatch]: {
+    connectorId: batchConnectorId,
+    operationId: sendtobatch,
   },
   [compose]: {
     connectorId: dataOperationConnectorId,
@@ -532,10 +528,6 @@ const builtInOperationsMetadata: Record<string, OperationInfo> = {
   [select]: {
     connectorId: dataOperationConnectorId,
     operationId: select,
-  },
-  [sendtobatch]: {
-    connectorId: batchConnectorId,
-    operationId: sendtobatch,
   },
   [setvariable]: {
     connectorId: variableConnectorId,